Put calendar in component
